telegram_bot/requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.5
redis>=5.0.0
pydantic>=2.0.0 